class_agents/pyproject.toml
[project]
name = "class-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
